Fixed the circular dependency injection
