+++ src/WordGame/LocalDictionaryHelpers/ITemporaryDefinitionHelper.cs
@@ -1,0 +1,7 @@
+﻿namespace Chat.WordGame.LocalDictionaryHelpers
+{
+    public interface ITemporaryDefinitionHelper
+    {
+        void AutomaticallySetTemporaryDefinitionForWord(string filename, string word, string temporaryDefinition);
+    }
+}
