+++ src/WordGame/WordHelpers/IWordService.cs
@@ -8,7 +8,6 @@ namespace Chat.WordGame.WordHelpers
         string GetDefinition(string filename, string word);
         void AddNewWordToDictionary(string filename, string word, string definition);
         void UpdateExistingWord(string filename, string word, string definition);
-        void AutomaticallySetTemporaryDefinitionForWord(string filename, string word, string temporaryDefinition);
         void ToggleIsWordInDictionary(string filename, string word);
     }
 }
